Conflicts:
	modules.git
	modules.svn
	pack.sh
	pack_all.sh
	tag_all.sh
	version

+++ pack_l10n.sh
@@ -2,15 +2,14 @@
 
 . version
 . utils.sh
+. config
 
 force=$1
-
-branch=svn://anonsvn.kde.org/home/kde/branches/stable/l10n-kde4
 
 mkdir -p sources/kde-l10n
 mkdir -p versions
 
-svn cat $branch/scripts/autogen.sh > /tmp/kde-l10n-autogen.sh
+svn cat $l10n_repo/scripts/autogen.sh > /tmp/kde-l10n-autogen.sh
 chmod +x /tmp/kde-l10n-autogen.sh
 
 remove_stuff()
@@ -69,7 +68,7 @@ pack_lang()
         # the rev of the root l10n dir and thus the checkDownloadUptodate thing
         # works for langs inside langs (like sr)
         rev=`get_svn_rev`
-        svn export $branch/$lang@ $lang@ &> /dev/null
+        svn export $l10n_repo/$lang@ $lang@ &> /dev/null
         rev2=`get_svn_rev`
         if [ "$rev" = "$rev2" ]; then
             if [ $rootLang -eq 1 ]; then
@@ -79,23 +78,28 @@ pack_lang()
             cd $lang
             remove_stuff
             pack_variants
-            cd ..
-            if [ $rootLang -eq 1 ]; then
-                # Delete empty folders, we do it a few times
-                # in case there is empty dirs inside empty dirs
-                find $lang -type d -empty -delete
-                find $lang -type d -empty -delete
-                find $lang -type d -empty -delete
-                find $lang -type d -empty -delete
-                /tmp/kde-l10n-autogen.sh $lang
-                mv $lang kde-l10n-$lang-$version
-                find kde-l10n-$lang-$version -type f |sed 's/^\.*\/*//'|sort > MANIFEST
-                tar cf kde-l10n-$lang-$version.tar --owner 0 --group 0 --numeric-owner --no-recursion --files-from MANIFEST
-                xz -9 kde-l10n-$lang-$version.tar
-                mv kde-l10n-$lang-$version.tar.xz sources/kde-l10n
-                rm -f MANIFEST
+            rev3=`get_svn_rev`
+            # Check again after pack_variants just in case
+            # something happened in between
+            if [ "$rev" = "$rev3" ]; then
+                cd ..
+                if [ $rootLang -eq 1 ]; then
+                    # Delete empty folders, we do it a few times
+                    # in case there is empty dirs inside empty dirs
+                    find $lang -type d -empty -delete
+                    find $lang -type d -empty -delete
+                    find $lang -type d -empty -delete
+                    find $lang -type d -empty -delete
+                    /tmp/kde-l10n-autogen.sh $lang
+                    mv $lang kde-l10n-$lang-$version
+                    find kde-l10n-$lang-$version -type f |sed 's/^\.*\/*//'|sort > MANIFEST
+                    tar cf kde-l10n-$lang-$version.tar --owner 0 --group 0 --numeric-owner --no-recursion --files-from MANIFEST
+                    xz -9 kde-l10n-$lang-$version.tar
+                    mv kde-l10n-$lang-$version.tar.xz sources/kde-l10n
+                    rm -f MANIFEST
+                fi
+                checkout=0
             fi
-            checkout=0
         fi
         if [ $rootLang -eq 1 ]; then
             rm -rf $lang kde-l10n-$lang-$version
@@ -106,7 +110,7 @@ pack_lang()
 cat language_list | while read lang; do
     finalDestination=sources/kde-l10n/kde-l10n-$lang-$version.tar.xz
     versionFilePath=versions/kde-l10n-$lang
-    repoLine="$branch/$lang"
+    repoLine="$l10n_repo/$lang"
 
     checkDownloadUptodate "svn"
     uptodate=$?
